ml-for-bem/sampling-requirements.txt
numpy
requests
google-cloud-storage
boto3
tables
geopandas<0.13
pandas<2
shapely<2
pandas<2
h5py
fiona
pvlib
ladybug-core
python-dotenv
jsondiff